requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
pillow
djangorestframework-simplejwt
dj-config-url
django-filter
drf-yasg
django-star-ratings